change openpgpexchange signature

+++ stm32l432/openpgplib.cpp
@@ -13,7 +13,7 @@
 #include "applets/apduconst.h"
 
 Applet::APDUExecutor *fexecutor = nullptr;
-void OpenpgpExchange(uint8_t *datain, size_t datainlen, uint8_t *dataout, size_t *outlen) {
+void OpenpgpExchange(uint8_t *datain, size_t datainlen, uint8_t *dataout, uint32_t *outlen) {
 	*outlen = 0;
 
 	if (fexecutor == nullptr)
